More work on disagg_by_src [skip hazardlib]
